fix: refactor code and add underlines for active links
