+++ src/components/navigation/navbar-link.tsx
@@ -22,16 +22,15 @@ export function NavbarLink({
     <Link
       href={`/${path}`}
       className={cn(
-        "text-background underline-offset-4 transition-colors duration-300 lg:border-l-2 lg:px-[23px] lg:text-xl lg:no-underline lg:last:border-r-2 xl:px-8",
+        "text-background underline-offset-4 transition-colors duration-300 lg:border-l-2 lg:px-[23px] lg:text-xl lg:last:border-r-2 xl:px-8",
         {
-          "lg:text-primary sm:hover:text-secondary/80 underline":
+          "lg:text-primary sm:hover:text-secondary/80 lg:hover:text-primary/40 border-primary underline lg:no-underline":
             route === path && !isFixed,
-          "text-secondary hover:text-secondary/80 underline":
-            route === path && isFixed,
-          "text-primary-foreground hover:text-secondary/80 border-secondary/80 text-lg":
-            isFixed,
-          "lg:text-foreground lg:hover:text-primary/85 hover:text-secondary/80 border-primary text-lg":
-            !isFixed,
+          "hover:text-secondary/80 underline": route === path && isFixed,
+          "text-primary-foreground hover:text-secondary/80 border-secondary/80":
+            isFixed && route !== path,
+          "lg:text-foreground lg:hover:text-primary/85 hover:text-secondary/80 border-primary":
+            !isFixed && route !== path,
         },
       )}
     >
