+++ src/hooks/use-scroll.tsx
@@ -1,0 +1,19 @@
+import { useEffect, useState } from "react";
+
+export function useScroll() {
+  const [scrollPosition, setScrollPosition] = useState({ x: 0, y: 0 });
+
+  useEffect(() => {
+    const handleScroll = () => {
+      setScrollPosition({ x: window.scrollX, y: window.scrollY });
+    };
+
+    handleScroll();
+    window.addEventListener("scroll", handleScroll);
+    return () => {
+      window.removeEventListener("scroll", handleScroll);
+    };
+  }, []);
+
+  return scrollPosition;
+}
